+++ app/src/main/java/org/sopt/do_sopt_compose/MainActivity.kt
@@ -16,12 +16,11 @@ import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
+import org.sopt.do_sopt_compose.navigation.AuthNavigation
+import org.sopt.do_sopt_compose.navigation.Screen
 import org.sopt.do_sopt_compose.ui.pages.LoginPage
-import org.sopt.do_sopt_compose.ui.pages.states.LoginPageState
 import org.sopt.do_sopt_compose.ui.pages.MainPage
 import org.sopt.do_sopt_compose.ui.pages.SignUpPage
-import org.sopt.do_sopt_compose.ui.pages.states.MainPageState
-import org.sopt.do_sopt_compose.ui.pages.states.SignUpPageState
 import org.sopt.do_sopt_compose.ui.theme.Do_sopt_composeTheme
 
 class MainActivity : ComponentActivity() {
@@ -35,55 +34,10 @@ class MainActivity : ComponentActivity() {
                 ) {
                     Box(modifier = Modifier.fillMaxSize()) {
                         val navController = rememberNavController()
-                        NavHost(navController, startDestination = Screen.Login.route) {
-                            addLogin(navController = navController)
-                            addSignUp(navController = navController)
-                            addMain(navController = navController)
-                        }
+                        AuthNavigation(navController = navController)
                     }
                 }
             }
         }
-    }
-}
-
-private fun NavGraphBuilder.addLogin(navController: NavController) {
-    composable(route = Screen.Login.route) {
-        // 버튼 클릭 시 화면 이동
-        LoginPage(
-            navController = navController,
-        )
-    }
-}
-
-private fun NavGraphBuilder.addSignUp(navController: NavController) {
-    composable(route = Screen.SignUp.route) {
-        SignUpPage(
-            navController = navController,
-        )
-    }
-}
-
-private fun NavGraphBuilder.addMain(navController: NavController) {
-    composable(route = Screen.Main.route) {
-        MainPage(
-            navController = navController,
-        )
-    }
-}
-
-@Composable
-fun Greeting(name: String, modifier: Modifier = Modifier) {
-    Text(
-        text = "Hello $name!",
-        modifier = modifier,
-    )
-}
-
-@Preview(showBackground = true)
-@Composable
-fun GreetingPreview() {
-    Do_sopt_composeTheme {
-        Greeting("Android")
     }
 }
