+++ app/src/main/java/org/sopt/do_sopt_compose/navigation/Screen.kt
@@ -1,4 +1,4 @@
-package org.sopt.do_sopt_compose
+package org.sopt.do_sopt_compose.navigation
 
 sealed class Screen(val route: String) {
     object Login : Screen(route = "login")
